+++ extlib/benz/macro.c
@@ -159,7 +159,7 @@ pic_macro_variable_p(pic_state *pic)
 static pic_value
 pic_macro_variable_eq_p(pic_state *pic)
 {
-  size_t argc, i;
+  int argc, i;
   pic_value *argv;
 
   pic_get_args(pic, "*", &argc, &argv);
